# CAB230 Server Side Application/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Movie API - n10869069",
        "description": "This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below.",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "n10869069@qut.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter JWT Bearer token **_only_**"
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/user/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "description": "Generates a new access token using a valid refresh token.",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "example": "ajsonwebtoken"
                                    }
                                },
                                "required": [
                                    "refreshToken"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token successfully refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "bearerToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "refreshToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "bearerToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Bearer",
                                        "expires_in": 600
                                    },
                                    "refreshToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Refresh",
                                        "expires_in": 86400
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid refresh request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ajsonwebtoken"
                                        }
                                    }
                                },
                                "example": {
                                    "error": true,
                                    "message": "Request body incomplete, refresh token required"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Click on 'Schema' below to see the possible error responses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "JWT token has expired"
                                        }
                                    }
                                }
                            }
                        }
                    }
                    
                }
            }
        },
...
    }
}